FUSE/requirements-fuse2rest.txt
psutil
fusepy
msgpack

requests  # temporary...
